Bundler fixes w.r.t. importmap
